Attempt to fix pytube error when downloading videos
Currently, when clicking the download button and passing all validations, pytube throws the error "regex_search: could not find match for (?:v=|/)([0-9A-Za-z_-]{11}).*". This error seems to be related to an issue with the pytube library. After researching the issue and attempting to update the pytube version and fix any issues with the code, the error persists. This commit serves as an attempt to fix the error and continue working towards a solution.