chore: Update app version to 2.1.0
In this commit, I've updated the version of the application in the metadata files. The new version is 2.1.0, reflecting the latest changes and improvements made to the app.

By keeping the app version up to date, we ensure that users have the latest and most stable version of the application available.